refactor(MenuGroup) : 메뉴그룹 리팩토링
- 인수테스트
- 테스트 픽스처
- JDBCTemplate에서 JPA 로 변경
