+++ src/main/java/kitchenpos/menu/domain/MenuGroupDao.java
@@ -1,0 +1,6 @@
+package kitchenpos.menu.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuGroupDao extends JpaRepository<MenuGroup, Long> {
+}
