+++ src/main/java/kitchenpos/menu/dto/MenuGroupRequest.java
@@ -1,16 +1,10 @@
 package kitchenpos.menu.dto;
 
 public class MenuGroupRequest {
-    private Long id;
     private String name;
 
-    public MenuGroupRequest(Long id, String name) {
-        this.id = id;
+    public MenuGroupRequest(String name) {
         this.name = name;
-    }
-
-    public Long getId() {
-        return id;
     }
 
     public String getName() {
