+++ src/test/java/common/MenuGroupFixture.java
@@ -4,31 +4,16 @@ import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuGroupFixture {
 
-
     public static MenuGroup 메뉴그룹_두마리() {
-
-        MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setId(1L);
-        menuGroup.setName("두마리메뉴");
-        return menuGroup;
+        return MenuGroup.of(1L, "두마리메뉴");
     }
 
     public static MenuGroup 메뉴그룹_한마리() {
-
-        MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setId(2L);
-        menuGroup.setName("한마리메뉴");
-
-        return menuGroup;
+        return MenuGroup.of(2L, "한마리메뉴");
     }
 
     public static MenuGroup 메뉴그룹_신메뉴() {
-
-        MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setId(3L);
-        menuGroup.setName("신메뉴");
-
-        return menuGroup;
+        return MenuGroup.of(3L, "신메뉴");
     }
 
 }
